% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dsproject.R
\name{create_dsproject}
\alias{create_dsproject}
\title{Creates a template of folders and files for the 'ideal' data science project}
\usage{
create_dsproject(path, open = FALSE)
}
\arguments{
\item{path}{A path where to create the project template. Can be relative, absolute and non existent.}

\item{open}{whether to open the RStudio project or not. Set to
FALSE by default}
}
\value{
Nothing, it creates and edits several folders and files in \code{path}
}
\description{
Creates a template of folders and files for the 'ideal' data science project
}
\details{
The function accepts a valid path (either relative or absolute) and applies these steps:

text describing parameter inputs in more detail.
\itemize{
 \item{"Folders"}{Creates folders code, data, report and misc inside \code{path}}
 \item{"RProjects"}{Creates an R project in \code{path}}
 \item{"Git"}{Initializes a Git repository in \code{path}}
 \item{"Documentation"}{Adds a README.Rmd for project purposes}
 \item{"Package dependency"}{Installs and loads \code{packrat} for package dependency management}
 \item{"Fresh start"}{Restarts R and opens the newly created .Rproj with packrat loaded}
}
}
\examples{

\dontrun{
create_dsproject()
}

}
